+++ security/provenance/include/provenance_filter.h
@@ -61,8 +61,6 @@ static __always_inline bool filter_update_node(const uint64_t relation_type)
 		return true;
 	if (relation_type == RL_NAMED)
 		return true;
-	if (relation_type == RL_NAMED_PROCESS)
-		return true;
 	return false;
 }